refactor: move user-realted api action to subfolder
