+++ project/src/store/user/user.slice.ts
@@ -2,7 +2,7 @@ import { PayloadAction, createSlice } from '@reduxjs/toolkit';
 import { AuthorizationStatus } from 'src/consts/api';
 import { DomainNamespace } from 'src/consts/domain';
 import { UserLogin } from 'src/types/types';
-import { checkIfUserAuthorizedAction, logUserInAction, logUserOutAction } from '../api-actions';
+import { checkIfUserAuthorizedAction, logUserInAction, logUserOutAction } from 'src/store/api-user/api-user.actions';
 import { dropToken, setToken } from 'src/services/token';
 
 const initialState = {
